% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplify.R
\name{samplify}
\alias{samplify}
\title{Created a resampled tibble}
\usage{
samplify(data, times, size, ..., replace = FALSE, key = ".sample")
}
\arguments{
\item{data}{A tbl.}

\item{times}{A single integer specifying the number of resamples.
If the \code{tibble} is grouped, this is the number of resamples per group.}

\item{size}{A single integer specifying the size of each resample.}

\item{...}{Not used.}

\item{replace}{Whether or not to sample with replacement.}

\item{key}{A single character specifying the name of the virtual group
that is added.}
}
\value{
A \code{resampled_df} with an extra group specified by \code{key}.
}
\description{
\code{samplify()} creates a resampled tibble with \emph{virtual groups}.
}
\details{
Currently you can use \code{\link[dplyr:summarise]{dplyr::summarise()}}, \code{\link[dplyr:do]{dplyr::do()}}, and the group
manipulation functions: \code{\link[dplyr:group_map]{dplyr::group_map()}}, \code{\link[dplyr:group_walk]{dplyr::group_walk()}},
\code{\link[dplyr:group_nest]{dplyr::group_nest()}} and \code{\link[dplyr:group_split]{dplyr::group_split()}} on a \code{resampled_df}.
}
\examples{
library(dplyr)
library(broom)

samplify(iris, times = 3, size = 20)

iris \%>\%
  samplify(times = 3, size = 20) \%>\%
  summarise(per_strap_mean = mean(Petal.Width))

iris \%>\%
  group_by(Species) \%>\%
  samplify(times = 3, size = 20) \%>\%
  summarise(per_strap_species_mean = mean(Petal.Width))

# Alter the name of the group with `key`
# Materialize them with collect()
samps <- samplify(iris, times = 3, size = 5, key = ".samps")
collect(samps)

collect(samps, id = ".id", original_id = ".orig_id")

}
\seealso{
Other virtual samplers: \code{\link{bootstrapify}}
}
\concept{virtual samplers}
